+++ apidocs/interfaces/callresult.md
@@ -1,0 +1,70 @@
+[wasm-imagemagick](../README.md) > [CallResult](../interfaces/callresult.md)
+
+# Interface: CallResult
+
+The result of calling [call](../#call). Also the base class of results of calling [execute](executioncontext.md#execute).
+
+## Hierarchy
+
+**CallResult**
+
+↳  [ExecuteResultOne](executeresultone.md)
+
+## Index
+
+### Properties
+
+* [exitCode](callresult.md#exitcode)
+* [outputFiles](callresult.md#outputfiles)
+* [stderr](callresult.md#stderr)
+* [stdout](callresult.md#stdout)
+
+---
+
+## Properties
+
+<a id="exitcode"></a>
+
+###  exitCode
+
+**● exitCode**: *`number`*
+
+*Defined in [magickApi.ts:51](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L51)*
+
+Exit code of the command executed. If 0 the command executed successfully, otherwise an error occurred and `stderr` could have some information about what was wrong
+
+___
+<a id="outputfiles"></a>
+
+###  outputFiles
+
+**● outputFiles**: *[MagickOutputFile](magickoutputfile.md)[]*
+
+*Defined in [magickApi.ts:39](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L39)*
+
+Output files generated by the command, if any
+
+___
+<a id="stderr"></a>
+
+###  stderr
+
+**● stderr**: *`string`[]*
+
+*Defined in [magickApi.ts:47](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L47)*
+
+Output printed by the command to stderr. If `exitCode != 0` then this property could have some information about the error.
+
+___
+<a id="stdout"></a>
+
+###  stdout
+
+**● stdout**: *`string`[]*
+
+*Defined in [magickApi.ts:43](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L43)*
+
+Output printed by the command to stdout. For example the command `identify rose:` will print useful information to stdout
+
+___
+
