+++ apidocs/interfaces/executeconfig.md
@@ -1,0 +1,38 @@
+[wasm-imagemagick](../README.md) > [ExecuteConfig](../interfaces/executeconfig.md)
+
+# Interface: ExecuteConfig
+
+## Hierarchy
+
+**ExecuteConfig**
+
+## Index
+
+### Properties
+
+* [commands](executeconfig.md#commands)
+* [inputFiles](executeconfig.md#inputfiles)
+
+---
+
+## Properties
+
+<a id="commands"></a>
+
+###  commands
+
+**● commands**: *[ExecuteCommand](../#executecommand)*
+
+*Defined in [execute.ts:13](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L13)*
+
+___
+<a id="inputfiles"></a>
+
+### `<Optional>` inputFiles
+
+**● inputFiles**: *[MagickInputFile](magickinputfile.md)[]*
+
+*Defined in [execute.ts:10](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L10)*
+
+___
+
