+++ apidocs/interfaces/executeevent.md
@@ -1,0 +1,48 @@
+[wasm-imagemagick](../README.md) > [ExecuteEvent](../interfaces/executeevent.md)
+
+# Interface: ExecuteEvent
+
+## Hierarchy
+
+**ExecuteEvent**
+
+## Index
+
+### Properties
+
+* [command](executeevent.md#command)
+* [id](executeevent.md#id)
+* [took](executeevent.md#took)
+
+---
+
+## Properties
+
+<a id="command"></a>
+
+###  command
+
+**● command**: *[Command](../#command)*
+
+*Defined in [execute.ts:113](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L113)*
+
+___
+<a id="id"></a>
+
+###  id
+
+**● id**: *`number`*
+
+*Defined in [execute.ts:115](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L115)*
+
+___
+<a id="took"></a>
+
+###  took
+
+**● took**: *`number`*
+
+*Defined in [execute.ts:114](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L114)*
+
+___
+
