+++ apidocs/interfaces/executelistener.md
@@ -1,0 +1,54 @@
+[wasm-imagemagick](../README.md) > [ExecuteListener](../interfaces/executelistener.md)
+
+# Interface: ExecuteListener
+
+## Hierarchy
+
+**ExecuteListener**
+
+## Index
+
+### Methods
+
+* [afterExecute](executelistener.md#afterexecute)
+* [beforeExecute](executelistener.md#beforeexecute)
+
+---
+
+## Methods
+
+<a id="afterexecute"></a>
+
+### `<Optional>` afterExecute
+
+▸ **afterExecute**(event: *[ExecuteEvent](executeevent.md)*): `void`
+
+*Defined in [execute.ts:119](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L119)*
+
+**Parameters:**
+
+| Name | Type |
+| ------ | ------ |
+| event | [ExecuteEvent](executeevent.md) |
+
+**Returns:** `void`
+
+___
+<a id="beforeexecute"></a>
+
+### `<Optional>` beforeExecute
+
+▸ **beforeExecute**(event: *[ExecuteEvent](executeevent.md)*): `void`
+
+*Defined in [execute.ts:120](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L120)*
+
+**Parameters:**
+
+| Name | Type |
+| ------ | ------ |
+| event | [ExecuteEvent](executeevent.md) |
+
+**Returns:** `void`
+
+___
+
