+++ apidocs/interfaces/executeresultone.md
@@ -1,0 +1,86 @@
+[wasm-imagemagick](../README.md) > [ExecuteResultOne](../interfaces/executeresultone.md)
+
+# Interface: ExecuteResultOne
+
+## Hierarchy
+
+ [CallResult](callresult.md)
+
+**↳ ExecuteResultOne**
+
+↳  [ExecuteResult](executeresult.md)
+
+## Index
+
+### Properties
+
+* [errors](executeresultone.md#errors)
+* [exitCode](executeresultone.md#exitcode)
+* [outputFiles](executeresultone.md#outputfiles)
+* [stderr](executeresultone.md#stderr)
+* [stdout](executeresultone.md#stdout)
+
+---
+
+## Properties
+
+<a id="errors"></a>
+
+###  errors
+
+**● errors**: *`any`[]*
+
+*Defined in [execute.ts:52](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L52)*
+
+___
+<a id="exitcode"></a>
+
+###  exitCode
+
+**● exitCode**: *`number`*
+
+*Overrides [CallResult](callresult.md).[exitCode](callresult.md#exitcode)*
+
+*Defined in [execute.ts:53](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/execute.ts#L53)*
+
+___
+<a id="outputfiles"></a>
+
+###  outputFiles
+
+**● outputFiles**: *[MagickOutputFile](magickoutputfile.md)[]*
+
+*Inherited from [CallResult](callresult.md).[outputFiles](callresult.md#outputfiles)*
+
+*Defined in [magickApi.ts:39](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L39)*
+
+Output files generated by the command, if any
+
+___
+<a id="stderr"></a>
+
+###  stderr
+
+**● stderr**: *`string`[]*
+
+*Inherited from [CallResult](callresult.md).[stderr](callresult.md#stderr)*
+
+*Defined in [magickApi.ts:47](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L47)*
+
+Output printed by the command to stderr. If `exitCode != 0` then this property could have some information about the error.
+
+___
+<a id="stdout"></a>
+
+###  stdout
+
+**● stdout**: *`string`[]*
+
+*Inherited from [CallResult](callresult.md).[stdout](callresult.md#stdout)*
+
+*Defined in [magickApi.ts:43](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L43)*
+
+Output printed by the command to stdout. For example the command `identify rose:` will print useful information to stdout
+
+___
+
