+++ apidocs/interfaces/magickfile.md
@@ -1,0 +1,34 @@
+[wasm-imagemagick](../README.md) > [MagickFile](../interfaces/magickfile.md)
+
+# Interface: MagickFile
+
+Base class for ImageMagick input and output files.
+
+## Hierarchy
+
+**MagickFile**
+
+↳  [MagickOutputFile](magickoutputfile.md)
+
+↳  [MagickInputFile](magickinputfile.md)
+
+## Index
+
+### Properties
+
+* [name](magickfile.md#name)
+
+---
+
+## Properties
+
+<a id="name"></a>
+
+###  name
+
+**● name**: *`string`*
+
+*Defined in [magickApi.ts:5](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L5)*
+
+___
+
