+++ apidocs/interfaces/magickinputfile.md
@@ -1,0 +1,46 @@
+[wasm-imagemagick](../README.md) > [MagickInputFile](../interfaces/magickinputfile.md)
+
+# Interface: MagickInputFile
+
+Represents input files that need to be provided to [call](../#call) or [execute](executioncontext.md#execute).
+
+Can be builded using [buildInputFile](../#buildinputfile)
+
+## Hierarchy
+
+ [MagickFile](magickfile.md)
+
+**↳ MagickInputFile**
+
+## Index
+
+### Properties
+
+* [content](magickinputfile.md#content)
+* [name](magickinputfile.md#name)
+
+---
+
+## Properties
+
+<a id="content"></a>
+
+###  content
+
+**● content**: *`Uint8Array`*
+
+*Defined in [magickApi.ts:21](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L21)*
+
+___
+<a id="name"></a>
+
+###  name
+
+**● name**: *`string`*
+
+*Inherited from [MagickFile](magickfile.md).[name](magickfile.md#name)*
+
+*Defined in [magickApi.ts:5](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L5)*
+
+___
+
