+++ apidocs/interfaces/magickoutputfile.md
@@ -1,0 +1,44 @@
+[wasm-imagemagick](../README.md) > [MagickOutputFile](../interfaces/magickoutputfile.md)
+
+# Interface: MagickOutputFile
+
+Represents output files generated when an ImageMagick command executes.
+
+## Hierarchy
+
+ [MagickFile](magickfile.md)
+
+**↳ MagickOutputFile**
+
+## Index
+
+### Properties
+
+* [blob](magickoutputfile.md#blob)
+* [name](magickoutputfile.md#name)
+
+---
+
+## Properties
+
+<a id="blob"></a>
+
+###  blob
+
+**● blob**: *`Blob`*
+
+*Defined in [magickApi.ts:12](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L12)*
+
+___
+<a id="name"></a>
+
+###  name
+
+**● name**: *`string`*
+
+*Inherited from [MagickFile](magickfile.md).[name](magickfile.md#name)*
+
+*Defined in [magickApi.ts:5](https://github.com/KnicKnic/WASM-ImageMagick/blob/7684a1c/src/magickApi.ts#L5)*
+
+___
+
