+++ spec/util/fileSpec.ts
@@ -1,4 +1,4 @@
-import { blobToString, buildInputFile, Call, compare, extractInfo, getFileNameExtension, inputFileToOutputFile, outputFileToInputFile, getFileName } from '../../src'
+import { blobToString, buildInputFile, Call, compare, extractInfo, getFileNameExtension, getFileName, asInputFile, asOutputFile } from '../../src'
 
 export default describe('util/file', () => {
 
@@ -36,22 +36,22 @@ export default describe('util/file', () => {
     xit('should support data:// urls with embedded image content', () => { })
   })
 
-  describe('outputFileToInputFile and inputFileToOutputFile', () => {
+  describe('asInputFile and asOutputFile', () => {
 
-    it('outputFileToInputFile should help to use output images in next commands as input images', async done => {
+    it('asInputFile should help to use output images in next commands as input images', async done => {
       const img = await buildInputFile('holocaust.jpg')
       const result1 = await Call([img], ['convert', 'holocaust.jpg', '-resize', '25%', 'holocaust.gif'])
       const info1 = await extractInfo(result1[0])
       expect(info1[0].image.geometry.width).toBe(80)
-      const result2 = await Call([await outputFileToInputFile(result1[0])], ['convert', 'holocaust.gif', '-resize', '40%', 'resized.png'])
+      const result2 = await Call([await asInputFile(result1[0])], ['convert', 'holocaust.gif', '-resize', '40%', 'resized.png'])
       const info2 = await extractInfo(result2[0])
       expect(info2[0].image.geometry.width).toBe(32)
       done()
     })
 
-    it('outputFileToInputFile and inputFileToOutputFile should generate equal images', async done => {
+    it('asInputFile and asOutputFile should generate equal images', async done => {
       const img = await buildInputFile('holocaust.jpg')
-      const img2 = await outputFileToInputFile(await inputFileToOutputFile(img), 'img2.jpg')
+      const img2 = await asInputFile(await asOutputFile(img), 'img2.jpg')
       expect(await compare(img, img2)).toBe(true)
       done()
     })
