+++ src/execute.ts
@@ -1,27 +1,61 @@
-import { MagickInputFile, MagickOutputFile, outputFileToInputFile, call, asCommand } from '.'
+import { MagickInputFile, MagickOutputFile, call, asCommand } from '.'
 import pMap from 'p-map'
 import { CallResult } from './magickApi'
 import { values } from './util/misc'
+import { asInputFile } from './util';
 
 export type Command = (string | number)[]
 
 export interface ExecuteConfig {
   inputFiles?: MagickInputFile[]
-  /** commands could be array form like [['convert', 'foo.png', 'bar.gif']] or CLI form like ['convert foo.png bar.gif'] */
+  /**
+   */
   commands: ExecuteCommand
 }
 
+/**
+ *
+ * Commands could have the following syntaxes:
+ *  * array form like `[['convert', 'foo.png', 'bar.gif'], ['identify', 'bar.gif']]`
+ *  * just one array: `['convert', 'foo.png', 'bar.gif']`
+ *  * command line strings: `['convert foo.png bar.gif', 'idenfity bar.gif']`
+ *  * just one string: `'convert foo.png bar.gif'`
+ *
+ * Also, for command line strings, multiple commands can be specified in the same string separating with new lines:
+ *
+ * ```js
+ * const result = await execute(`
+ *   convert rose: -sharpen 0x1 reconstruct.jpg
+ *   compare rose: reconstruct.jpg difference.png
+ *   compare -compose src rose: reconstruct.jpg difference.png
+ * `)
+ * ```
+ *
+ * Also, command line strings support breaking the same command in multiple lines by using `\` like in:
+ *
+ * ```js
+ * const result = await execute(`
+ *   convert -size 250x100 xc: +noise Random -channel R -threshold .4% \\
+ *     -negate -channel RG -separate +channel \\
+ *     \( +clone \) -compose multiply -flatten \\
+ *     -virtual-pixel Tile -background Black \\
+ *     -blur 0x.6 -motion-blur 0x15-90 -normalize \\
+ *     +distort Polar 0 +repage 'star inward.gif'
+ * `)
+ * ```
+ *
+ * If you need to escape arguments like file names with spaces, use single quotes `'`, like the output file in the previous example `'star inward.gif'`
+ */
 export type ExecuteCommand = Command[] | string[] | string
 
 export interface ExecuteResultOne extends CallResult {
-  // outputFiles: MagickOutputFile[]
   errors: any[]
-  // stderr: string[]
-  // stdout: string[]
   exitCode: number
 }
 
-/** execute first command in given config */
+/**
+ * Execute first command in given config.
+ */
 export async function executeOne(configOrCommand: ExecuteConfig | ExecuteCommand): Promise<ExecuteResultOne> {
   const config = asExecuteConfig(configOrCommand)
   let result: CallResult = {
@@ -60,15 +94,17 @@ export function asExecuteConfig(arg: ExecuteConfig | ExecuteCommand): ExecuteCon
     commands: arg,
   }
 }
+
 /**
- * Assumes the command won't fail so returns one output file directly or undefined if it's not found (or error occurs)
- * @param configOrCommand
+ * `execute()` shortcut that useful for commands that return only one output file or when only one particular output file is relevant.
  * @param outputFileName optionally user can give the desired output file name
+ * @returns If `outputFileName` is given the file with that name, the first output file otherwise or undefined
+ * if no file match, or no output files where generated (like in an error).
  */
 export async function executeAndReturnOutputFile(configOrCommand: ExecuteConfig | ExecuteCommand, outputFileName?: string): Promise<MagickOutputFile | undefined> {
   const config = asExecuteConfig(configOrCommand)
   const result = await execute(config)
-  return outputFileName ? result.outputFiles.find(f => f.name === outputFileName) : result.outputFiles.length && result.outputFiles[0]
+  return outputFileName ? result.outputFiles.find(f => f.name === outputFileName) : (result.outputFiles.length && result.outputFiles[0] || undefined)
 }
 
 // execute event emitter
@@ -92,43 +128,32 @@ export function addExecuteListener(l: ExecuteListener) {
 
 export interface ExecuteResult extends ExecuteResultOne {
   results: ExecuteResultOne[]
-  breakOnError?: boolean
+  // breakOnError?: boolean
 }
 
 /**
  * Execute all commands in given config serially in order. Output files from a command become available as
- * input files in next commands. The execution result will contain all generated outputFiles. If same file name
- * is used later command output files will override previous ones. Example:
+ * input files in next commands. In the following example we execute two commands. Notice how the second one uses `image2.png` which was the output file of the first one:
  *
  * ```ts
- * const {outputFiles} = await execute({
- *  inputFiles: [await buildInputFile('fn.png', 'image1.png')],
- *  commands: [
- *    ['convert', 'image1.png', "-bordercolor", "#ffee44", "-background", "#eeff55", "+polaroid", "image2.png"],
- *    // heads up: next command uses "image2.png" which was the output of previous command:
- *    ["convert", "image2.png", "-fill", "#997711", "-tint", "55"],
- *  ]
+ * const { outputFiles, exitCode, stderr} = await execute({
+ *   inputFiles: [await buildInputFile('fn.png', 'image1.png')],
+ *   commands: `
+ *     convert image1.png -bordercolor #ffee44 -background #eeff55 +polaroid image2.png
+ *     convert image2.png -fill #997711 -tint 55 image3.jpg
+ * `
  * })
+ * if (exitCode) {
+ *   alert(`There was an error with the command: ${stderr.join('\n')}`)
+ * }
+ * else {
+ *   await loadImageElement(outputFiles.find(f => f.name==='image3.jpg'), document.getElementById('outputImage'))
+ * }
  * ```
  *
- * An alternative syntax with CLI-like strings is also supported:
+ * See {@link ExecuteCommand} for different command syntax supported.
  *
- * ```ts
- * const {outputFiles} = await execute({inputFiles: [], commands: [
- *   'convert rose: -rotate 70 image2.gif',
- *   'convert image2.gif -resize 33 image3.gif'
- * ] })
- * ```
- *
- * Or if it's only one command using just a string:
- *
- * ```ts
- * const {outputFiles} = await execute({inputFiles: [foo], commands: `convert 'my face image.png' \\( +clone -channel R -fx B \\) +swap -channel B -fx v.R bar.gif`})
- * ```
- *
- * Note: in string syntax you must use single quotes for CLI arguments that need so (like 'my face image.png'). no multiline with \ is supported.
- *
- * ```
+ * See {@link ExecuteResult} for details on the object returned
  */
 
 export async function execute(configOrCommand: ExecuteConfig | ExecuteCommand): Promise<ExecuteResult> {
@@ -155,7 +180,7 @@ export async function execute(configOrCommand: ExecuteConfig | ExecuteCommand): 
     allStderr = allStderr.concat(result.stderr || [])
     await pMap(result.outputFiles, async f => {
       allOutputFiles[f.name] = f
-      const inputFile = await outputFileToInputFile(f)
+      const inputFile = await asInputFile(f)
       allInputFiles[inputFile.name] = inputFile
     })
   }
