+++ src/executionContext.ts
@@ -1,19 +1,40 @@
 import { createImageHome, execute, ExecuteCommand, ExecuteConfig, ExecuteResult, ImageHome, MagickFile, MagickInputFile } from '.'
-import { getBuiltInImages } from './util'
+import { asExecuteConfig } from './execute'
 
 /**
- * Allow multiple execute() calls remembering previus execute() generated output files and previous given input files that
+ * Allow multiple execute() calls remembering previous execute() generated output files and previous given input files that
  * can be used as input files in next calls.
  */
 export interface ExecutionContext {
-  execute(configOrCommands: ExecuteConfig|ExecuteCommand|string): Promise<ExecuteResult>
-  /** programmatically add new files so they are available if following execute() calls */
+  /**
+   * This behaves almost the same as {@link execute}.
+   */
+  execute(configOrCommands: ExecuteConfig | ExecuteCommand | string): Promise<ExecuteResult>
+
+  /**
+   * Programmatically add new files so they are available if following `execute()` calls.
+   */
   addFiles(files: MagickFile[]): void
-  /** get all the files currently available in this context */
+
+  /**
+   * Get all the files currently available in this context.
+   */
   getAllFiles(): Promise<MagickInputFile[]>
-  // /** add built-in images like `rose:` to this execution context so they are present in getAllFiles() */
+
+  /**
+   * Add ImageMagick built-in images like `rose:`, `logo:`, etc to this execution context so they are present in `getAllFiles()`.
+   */
   addBuiltInImages(): Promise<void>
-  getFile(name: string): Promise<MagickInputFile>
+
+  /**
+   * Get a file by name or undefined if none.
+   */
+  getFile(name: string): Promise<MagickInputFile|undefined>
+
+  /**
+   * Remove files by name.
+   * @returns the files actually removed.
+   */
   removeFiles(names: string[]): MagickInputFile[]
 }
 
@@ -22,14 +43,13 @@ class ExecutionContextImpl implements ExecutionContext {
   constructor(private imageHome: ImageHome = createImageHome()) {
   }
 
-  async execute(configOrCommands: ExecuteConfig|ExecuteCommand|string): Promise<ExecuteResult> {
-    const config = asConfig(configOrCommands)
-    // debugger
+  async execute(configOrCommands: ExecuteConfig | ExecuteCommand | string): Promise<ExecuteResult> {
+    const config = asExecuteConfig(configOrCommands)
     config.inputFiles.forEach(f => {
       this.imageHome.register(f)
     })
     const inputFiles = await this.imageHome.getAll()
-    const result = await execute({...config, inputFiles})
+    const result = await execute({ ...config, inputFiles })
     result.outputFiles.forEach(f => {
       this.imageHome.register(f)
     })
@@ -63,14 +83,6 @@ class ExecutionContextImpl implements ExecutionContext {
     return new ExecutionContextImpl(inheritFrom && (inheritFrom as ExecutionContextImpl).imageHome)
   }
 
-}
-
-function asConfig(configOrCommands: ExecuteConfig|ExecuteCommand|string): ExecuteConfig {
-  if (typeof configOrCommands === 'string') {
-    return {inputFiles: [], commands: [configOrCommands]}
-  }
-  return (configOrCommands as ExecuteConfig).inputFiles ? (configOrCommands as ExecuteConfig) :
-  ({commands: configOrCommands as ExecuteCommand, inputFiles: ([] as MagickInputFile[])} as ExecuteConfig)
 }
 
 export function newExecutionContext(inheritFrom?: ExecutionContext): ExecutionContext {
